fix(laravel): error handler only on api routes
